Add git hook to prevent cross-module commits
This is intended for use with git-subtree. The scenario is that a
project is working with code that is maintained seperately externally
but for whatever reason (usuall "haet git submodules, rawr") the project
would like to keep only a single, monolithic git repository internally.
In order to not break external collaboration completely, git-subtree is
used to merge and split internal and external changes. Sanity check:
don't make a commit in the internal project that crosses more than one
submodule. This hook is an early revision of a sanity check to prevent
such atrocities.